requirements.txt
arrow==0.15.7
asgiref==3.2.10
boto3==1.14.12
botocore==1.17.12
certifi==2020.6.20
chardet==3.0.4
defusedxml==0.6.0
Django==3.0.7
django-allauth==0.42.0
django-cleanup==5.0.0
django-countries==6.1.2
django-crispy-forms==1.9.1
django-debug-toolbar==2.2
django-filter==2.3.0
django-storages==1.9.1
djangorestframework==3.11.0
docutils==0.15.2
idna==2.9
importlib-metadata==1.7.0
jmespath==0.10.0
Markdown==3.2.2
pkg-resources==0.0.0
Pillow==7.1.2
oauthlib==3.1.0
psycopg2-binary==2.8.5
PyJWT==1.7.1
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.1
stripe==2.48.0
twilio==6.43.0
urllib3==1.25.9
zipp==3.1.0
